hcintelligencelayer/requirements.txt
fastapi
uvicorn
pydantic
requests
google-genai
openai
azure-identity